+++ include/nfx/Gl.h
@@ -13,5 +13,6 @@
 #include "nfx/graphics/gl/core/framebuffers/Renderbuffer.h"
 #include "nfx/graphics/gl/core/shaders/ShaderProgram.h"
 #include "nfx/graphics/gl/core/textures/Texture2D.h"
+#include "nfx/graphics/gl/core/textures/TextureCube.h"
 #include "nfx/graphics/gl/core/Context.h"
 #include "nfx/graphics/gl/core/GlDefinitions.h"